refactor: relocate escapeHtml utilities
