+++ .claspignore
@@ -1,1 +1,1 @@
-shared/escapeHtml.js
+tests/escapeHtml.js
