+++ src/board.html
@@ -146,7 +146,7 @@
 
 
   <script>
-    <?!= include('shared/escapeHtml.html'); ?>
+    <?!= include('escapeHtml.html'); ?>
     const SCRIPT_URL = '<?!= typeof scriptUrl !== "undefined" ? scriptUrl.replace("/dev","/exec") : "" ?>';
     const teacherParam = '<?!= typeof teacher !== "undefined" ? teacher : "" ?>';
     const gradeParam = '<?!= typeof grade !== "undefined" ? grade : "" ?>';
