+++ src/class-select.html
@@ -36,7 +36,7 @@
     </main>
   </div>
   <script>
-  <?!= include('shared/escapeHtml.html'); ?>
+  <?!= include('escapeHtml.html'); ?>
   const SCRIPT_URL = '<?!= scriptUrl.replace("/dev", "/exec") ?>';
   const enrolledClasses = JSON.parse(sessionStorage.getItem('enrolledClasses') || '[]');
