+++ src/escapeHtml.gs
