+++ src/manage.html
@@ -447,7 +447,7 @@
   </script>
 
   <script>
-  <?!= include('shared/escapeHtml.html'); ?>
+  <?!= include('escapeHtml.html'); ?>
     // Lucide Icons を SVG に置き換える
     function renderIcons(scope) {
       // lucide.js の初期化を待つため、わずかに遅延させる
