+++ src/profile.html
@@ -46,7 +46,7 @@
   </main>
   <div id="trophyModal" class="fixed inset-0 bg-black/80 z-50 flex items-center justify-center p-4 hidden"></div>
   <script>
-    <?!= include('shared/escapeHtml.html'); ?>
+    <?!= include('escapeHtml.html'); ?>
     const allTrophies=[
       {id:'t1',name:'Beginner',icon:'award'},
       {id:'t2',name:'Explorer',icon:'compass'},
