+++ tests/LeaderboardDOM.test.js
@@ -1,5 +1,5 @@
 const { JSDOM } = require('jsdom');
-const escapeHtml = require('../src/shared/escapeHtml.js');
+const escapeHtml = require('./escapeHtml.js');
 
 function renderLeaderboard(data, doc) {
   const tbody = doc.querySelector('#leaderboardTable tbody');
