+++ tests/escapeHtml.js
